# Dashboard/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sharecycle API",
    "description": "API documentation for the Sharecycle backend application.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "order-controller"
        ],
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "order-controller"
        ],
        "summary": "Cancel a pending order",
        "description": "Allows a user to cancel their own order if it is still in the 'PENDING' state.\n\nUpon a successful request, the following changes occur:\n- The **Order Status** is updated from `PENDING` to `CANCELED`.\n- The associated PayPal payment authorization is **voided**, releasing the hold on the user's funds.\n",
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders/admin/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "operationId": "getAdminOrderDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "order-controller"
        ],
        "operationId": "updateOrderByAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "order-controller"
        ],
        "summary": "Cancel a pending order",
        "description": "Allows an admin to cancel an order if it is still in the 'PENDING' state.\n\nUpon a successful request, the following changes occur:\n- The **Order Status** is updated from `PENDING` to `CANCELED`.\n- The associated PayPal payment authorization is **voided**, releasing the hold on the user's funds.\n",
        "operationId": "cancelOrderByAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/orders/{orderId}/reject": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Reject a pending order (Bicycle Owner)",
        "description": "Allows the owner of a bicycle to reject a rental request.\n\nUpon a successful request, the following changes occur:\n- The **Order Status** is updated from `PENDING` to `REJECTED`.\n- The associated PayPal payment authorization is **voided**, releasing the hold on the renter's funds.\n",
        "operationId": "rejectOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}/approve": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Approve a pending order (Bicycle Owner)",
        "description": "Allows the owner of a bicycle to approve a rental request. This is a critical action that triggers multiple changes in the system.\n\nUpon a successful request, the following changes occur:\n- The **Order Status** is updated from `PENDING` to `APPROVED`.\n- The previously authorized PayPal payment is **captured**, transferring the funds.\n- The associated **Bicycle's availability date** is updated to the day after this order's end date.\n- Any other `PENDING` orders that overlap with this rental period for the same bicycle will be automatically **rejected**.\n",
        "operationId": "approveOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "iphoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "idCardNumber": {
            "type": "string"
          },
          "idCardImageUrl": {
            "type": "string"
          }
        }
      },
...
      "OrderRequestDTO": {
        "type": "object",
        "properties": {
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          }
        }
      },
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "bicycleId": {
            "type": "integer",
            "format": "int64"
          },
          "bikeName": {
            "type": "string"
          },
          "bikeImageUrl": {
            "type": "string"
          },
          "pricePerDay": {
            "type": "number",
            "format": "double"
          },
          "renter": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "owner": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          },
          "rentalDays": {
            "type": "integer",
            "format": "int64"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "orderStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELED"
            ]
          }
        }
      },
...
      "AdminOrderDetailDTO": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "bikeId": {
            "type": "integer",
            "format": "int64"
          },
          "bikeName": {
            "type": "string"
          },
          "bikeImageUrl": {
            "type": "string"
          },
          "bikePrice": {
            "type": "number",
            "format": "double"
          },
          "bikeCity": {
            "type": "string"
          },
          "bikeDate": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "customerProfileImage": {
            "type": "string"
          },
          "vendorName": {
            "type": "string"
          },
          "vendorId": {
            "type": "integer",
            "format": "int64"
          },
          "vendorProfileImage": {
            "type": "string"
          },
          "dateFrom": {
            "type": "string"
          },
          "dateTo": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}